pyproject.toml
[tool.poetry]
name = "fetch"
version = "0.1.0"
description = ""
authors = ["perdanamain <firman.fp123@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
numpy = "^2.1.3"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
